Add files via upload
I need help making this website better. I want the layout to look the same as it does but incorporate AI features such as a chatbot, onedrive capabilities, as well as the main issue of this being a file. Is it possible to make this a proper website? Feel free to alter the code and send it back or comment code to add and adjust in.